Crate type_metadata

Re-exports

pub use type_metadata_derive::Metadata;

pub use type_metadata_derive::TypeDef;

pub use type_metadata_derive::TypeId;

Modules

form

interner

Macros

tuple_meta_type

Structs

ClikeEnumVariant

EnumVariantStruct

EnumVariantTupleStruct

EnumVariantUnit

MetaType
A metatype abstraction.

NamedField

Namespace
Represents the namespace of a type definition.

Registry

TypeDefClikeEnum

TypeDefEnum

TypeDefStruct

TypeDefTupleStruct

TypeDefUnion

TypeIdArray

TypeIdCustom

TypeIdSlice

TypeIdTuple

UnnamedField

Enums

Builtin
This struct just exists for the purpose of better JSON output.

EnumVariant

NamespaceError
An error that may be encountered upon constructing namespaces.

TypeDef

TypeId

TypeIdPrimitive

Traits

HasTypeDef
Types implementing this trait can communicate their type structure.

HasTypeId
Implementors return their meta type identifiers.

IntoCompact

Metadata