Enum type_metadata::TypeDef

pub enum TypeDef<F: Form = MetaForm> {
    Builtin(Builtin),
    Struct(TypeDefStruct<F>),
    TupleStruct(TypeDefTupleStruct<F>),
    ClikeEnum(TypeDefClikeEnum<F>),
    Enum(TypeDefEnum<F>),
    Union(TypeDefUnion<F>),
}

A type definition represents the internal structure of a concrete type.

Variants

Builtin(Builtin)

A builtin type that has an implied and known internal structure.

Struct(TypeDefStruct<F>)

A struct with named fields.

TupleStruct(TypeDefTupleStruct<F>)

A tuple-struct with unnamed fields.

ClikeEnum(TypeDefClikeEnum<F>)

A C-like enum with simple named variants.

Enum(TypeDefEnum<F>)

A Rust enum with different kinds of variants.

Union(TypeDefUnion<F>)

An unsafe Rust union type.

Methods

impl TypeDef

pub fn builtin() -> Self

Preferred way to create a builtin type definition.