Trait type_metadata::HasTypeDef

pub trait HasTypeDef {
    fn type_def() -> TypeDef;
}

Types implementing this trait can communicate their type structure.

If the current type contains any other types, type_def would register their metadata into the given registry. For instance, <Option<MyStruct>>::type_def() would register MyStruct metadata. All implementation must register these contained types' metadata.

Required methods

fn type_def() -> TypeDef

Returns the type definition for Self type.

Loading content...