Crate type_metadata

Efficient and compact serialization of Rust types.

This library provides structures to easily retrieve compile-time type information at runtime and also to serialize this information in a compact form.

Registry

At the heart of its functionality is the Registry that acts as cache for known strings and types in order to efficiently deduplicate them and thus compactify the overall serialization.

Type Information

Information about types is split into two halfs.

The type identifier or TypeId is accessed through the HasTypeId trait.

The type definition or TypeDef is accessed throught the HasTypeDef trait.

Both traits shall be implemented for all types that are serializable. For this the library provides implementations for all commonly used Rust standard types and provides derive macros for simpler implementation of user provided custom types.

Compaction Forms

There is an uncompact form, called MetaForm that acts as a bridge from compile-time type information at runtime in order to easily retrieve all information needed to uniquely identify types. The compact form is retrieved by the IntoCompact trait and internally used by the Registry in order to convert the uncompact strings and types into their compact form.

Symbols and Namespaces

Since symbol names are often shared across type boundaries the Registry also deduplicates them. To differentiate two types sharing the same name namespaces are used. Commonly the namespace is equal to the one where the type has been defined in. For Rust prelude types such as Option and Result the root namespace (empty namespace) is used.

To use this library simply use the MetaForm initially with your own data structures and at best make them generic over the Form trait just as has been done in this crate with TypeId and TypeDef in order to go for a simple implementation of IntoCompact. Use a single instance of the Registry for compaction and provide this registry instance upon serialization. Done.

A usage example can be found in ink! here: https://github.com/paritytech/ink/blob/master/abi/src/specs.rs

Modules

form
Provides some form definitions.

interner
Interning data structure and associated symbol definitions.

Macros

tuple_meta_type
Takes a number of types and returns a vector that contains their respective MetaType instances.

Structs

ClikeEnumVariant
A C-like enum variant.

EnumVariantStruct
A struct enum variant with named fields.

EnumVariantTupleStruct
A tuple struct enum variant.

EnumVariantUnit
An unit struct enum variant.

MetaType
A metatype abstraction.

NamedField
A named field.

Namespace
Represents the namespace of a type definition.

Registry
The registry for compaction of type identifiers and definitions.

TypeDefClikeEnum
A C-like enum type.

TypeDefEnum
A Rust enum, aka tagged union.

TypeDefStruct
A Rust struct with named fields.

TypeDefTupleStruct
A tuple struct with unnamed fields.

TypeDefUnion
A union, aka untagged union, type definition.

TypeIdArray
An array type identifier.

TypeIdCustom
A type identifier for custom type definitions.

TypeIdSlice
A type identifier to refer to slice type definitions.

TypeIdTuple
A type identifier to refer to tuple types.

UnnamedField
An unnamed field from either a tuple-struct type or a tuple-struct variant.

Enums

Builtin
This struct just exists for the purpose of better JSON output.

EnumVariant
A Rust enum variant.

NamespaceError
An error that may be encountered upon constructing namespaces.

TypeDef
A type definition represents the internal structure of a concrete type.

TypeId
A type identifier.

TypeIdPrimitive
Identifies a primitive Rust type.

Traits

HasTypeDef
Types implementing this trait can communicate their type structure.

HasTypeId
Implementors return their meta type identifiers.

IntoCompact
Compacts the implementor using a registry.

Metadata
A super trait that shall be implemented by all types implementing HasTypeId and HasTypedef in order to more easily manage them.

Derive Macros

Metadata

TypeDef

TypeId